pyproject.toml
[project]
name = "voltage-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=5.5.2",
    "fastapi>=0.115.12",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.10.1",
    "pydantic>=2.11.3",
    "requests>=2.33.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "requests-mock>=1.12.1",
    "ruff>=0.15.8",
    "sqlfluff>=4.1.0",
    "sqlfluff-templater-dbt>=4.1.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that exercise the real Airflow/GCS/BigQuery pipeline path",
    "integration_heavy: marks the slower rerun/backfill pipeline tests that are opt-in by environment variable",
    "integration_failure: marks targeted freshness/failure-path pipeline checks that intentionally create stale or failing conditions and are opt-in by environment variable",
]

[tool.ruff]
target-version = "py311"
